[REFACTOR] Factory introduced
A factory for GqSetup are introduced. It should be the only source
of GqSetups besides deserialisation.
