+++ core/src/test/java/org/mammon/math/GqTest.java
@@ -6,17 +6,19 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
+import static org.mammon.math.GqSetupFactory.factory;
+
 public class GqTest {
 	private Group<Gq> group;
 
 	@Test(expected = IllegalArgumentException.class)
 	public void qShouldDividePminus1() {
-		new Gq(new GqSetup(12, 23));
+		new Gq(factory(12, 23).generateSetup());
 	}
 
 	@Before
 	public void createGroup() {
-		this.group = new Gq(new GqSetup(11, 23));
+		this.group = new Gq(factory(11, 23).generateSetup());
 	}
 
 	@Test
